Move workshop pipeline image generation to main script
